api/requirements.txt
fastapi
uvicorn
openai
pandas
python-multipart
psycopg2
datetime
boto3
BeautifulSoup4
beautifulSoup4
pandasai
bs4
requests
python-dotenv
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
cleo==2.0.1
colorama==0.4.6
crashtest==0.4.1
dataclasses-json==0.5.7
distlib==0.3.6
dulwich==0.21.3
exceptiongroup==1.1.1
filelock==3.11.0
frozenlist==1.3.3
greenlet==2.0.2
html5lib==1.1
httplib2==0.22.0
idna==3.4
importlib-metadata==6.3.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.2.3
jsonschema==4.17.3
keyring==23.13.1
langchain==0.0.135
lockfile==0.12.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.2
openapi-schema-pydantic==1.2.4
packaging==23.0
pandas==2.0.0
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==1.0.0
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.0
protobuf==4.22.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
pyodbc==4.0.35
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.3.0
python-dateutil==2.8.2
pytz==2023.3
pywin32-ctypes==0.2.0
PyYAML==6.0
rapidfuzz==2.15.1
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
shellingham==1.5.0.post1
six==1.16.0
SQLAlchemy==1.4.47
tabulate==0.9.0
tenacity==8.2.2
tomli==2.0.1
tomlkit==0.11.7
tqdm==4.65.0
trove-classifiers==2023.3.9
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
webencodings==0.5.1
yarl==1.8.2
zipp==3.15.0
